Rename wrapped environments.
