--- .github/workflows/tests_linters.yml
@@ -1,25 +1,0 @@
-name: Tests and Linters 🧪
-
-on: [ push, pull_request ]
-
-jobs:
-  tests-and-linters:
-    name: "Python ${{ matrix.python-version }} on ${{ matrix.os }}"
-    runs-on: "${{ matrix.os }}"
-    timeout-minutes: 10
-
-    strategy:
-      matrix:
-        python-version: ["3.9"]
-        os: [self-hosted]
-
-    steps:
-      - name: Checkout og-marl
-        uses: actions/checkout@v3
-      - uses: actions/setup-python@v4
-        with:
-            python-version: "${{ matrix.python-version }}"
-      - name: Install pre-commit dependencies 🔧
-        run: pip install pre-commit ruff mypy isort
-      - name: Run linters 🖌️
-        run: pre-commit run --all-files --verbose
