--- .pre-commit-config.yaml
@@ -1,71 +1,0 @@
-
-default_language_version:
-  python: python3
-default_stages: [commit]
-repos:
-  - repo: https://github.com/codespell-project/codespell
-    rev: v2.2.6
-    hooks:
-      - id: codespell
-        name: codespell
-        description: Checks for common misspellings in text files.
-        entry: codespell
-        language: python
-        types: [text]
-        args:
-          - "-L childs,olt"
-
-  - repo: https://github.com/astral-sh/ruff-pre-commit
-    # Ruff version.
-    rev: v0.2.1
-    hooks:
-      # Run the linter.
-      - id: ruff
-        args: [ --fix, --show-fixes ]
-      # Run the formatter.
-      - id: ruff-format
-
-  - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.1.0
-    hooks:
-      - id: end-of-file-fixer
-        name: "End of file fixer"
-      - id: debug-statements
-        name: "Debug imports and breakpoints checker"
-      - id: requirements-txt-fixer
-        name: "Requirements txt fixer"
-      - id: mixed-line-ending
-        name: "Mixed line ending fixer"
-      - id: check-yaml
-        name: "Yaml checker"
-        args: [ '--unsafe' ]
-      - id: trailing-whitespace
-        name: "Trailing whitespace fixer"
-
-  - repo: https://github.com/alessandrojcm/commitlint-pre-commit-hook
-    rev: v4.1.0
-    hooks:
-      - id: commitlint
-        name: "Commit linter"
-        stages: [ commit-msg ]
-        additional_dependencies: [ '@commitlint/config-conventional' ]
-
-  - repo: local
-    hooks:
-      - id: mypy
-        name: "Static type checker"
-        entry: python -m mypy og_marl
-        language: system
-        types: [ python ]
-        pass_filenames: false
-
-  # - repo: https://github.com/Lucas-C/pre-commit-hooks
-  #   rev: v1.3.0
-  #   hooks:
-  #     - id: insert-license
-  #       name: "License inserter"
-  #       files: .*py$
-  #       args:
-  #         - --license-filepath=docs/license_header.txt
-  #         - --comment-style=#
-  #       exclude: .npy$  # Do not add license to .npy files (the standard binary file format in NumPy)
