+++ og_marl/wrapped_environments/flatland_wrapper.py
@@ -23,7 +23,7 @@ from flatland.envs.rail_env import RailEnv
 from flatland.envs.rail_generators import sparse_rail_generator
 from gymnasium.spaces import Box, Discrete
 
-from og_marl.environment_wrappers.base import BaseEnvironment, Observations, ResetReturn, StepReturn
+from og_marl.wrapped_environments.base import BaseEnvironment, Observations, ResetReturn, StepReturn
 
 FLATLAND_MAP_CONFIGS = {
     "3trains": {
